XceTestAnalyze/requirements.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
HTMLParser==0.0.2
idna==2.9
jenkins==1.0.2
kiwisolver==1.2.0
matplotlib==3.2.2
multi-key-dict==2.0.3
mysql-connector-python==8.0.20
numpy==1.19.0
pandas==1.0.5
pbr==5.4.5
protobuf==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-jenkins==1.7.0
pytz==2020.1
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
urllib3==1.25.9

